Blueprint full implementation, added floors, changed attach behavior. Bugs: Odd values don't work, new attach method needs 2d implementation.

+++ core/src/com/prizmj/display/parts/BasicRoom.java
@@ -19,7 +19,7 @@ public class BasicRoom extends Room {
      * @param height - The height of the room. (approx.)
      * @param floorColor - The color the floor of the room.
      */
-    public BasicRoom(String name, float x, float y, float z, int width, int height, Color floorColor) {
+    public BasicRoom(String name, float x, float y, float z, float width, float height, Color floorColor) {
         super(name, x, y, z, width, height, floorColor);
     }
 }
